add all http methods + refacto entity class

+++ Java/spring-boot-rest-api/src/main/java/com/example/demo/repository/ProductRepository.java
@@ -7,13 +7,12 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 import javax.transaction.Transactional;
-import java.math.BigDecimal;
 
 @Repository
-public interface ProductRepository extends JpaRepository<Product, Integer> {
+public interface ProductRepository extends JpaRepository<Product, Long> {
 
     @Modifying
     @Transactional
     @Query("UPDATE product p SET p.description = ?1, p.category = ?2, p.price = ?3 WHERE p.id = ?4")
-    void updateById(String description, String category, BigDecimal price, int id);
+    void updateById(String description, String category, int price, long id);
 }
